Miscallaneous Cleanup (#5703)
* Cleanup - Unnecessary Boxing

* Cleanup - Unnecessary Unboxing

* Cleanup - For-Each Loops

* Cleanup - `indexOf != -1` -> `contains`

* Cleanup - Merge identical catch blocks

* Cleanup - Try-with-resources

* Cleanup - System.lineSeparator

* Cleanup - Reference types to primitives
Some loops over Integers were switched to IntStreams to hopefully cut down on overall boxing.

* Cleanup - Manually filling and copying arrays

* Remove unused imports

* Switch a lambda to a method reference

---------

+++ forge-adventure/src/main/java/forge/adventure/editor/ItemEdit.java
@@ -57,7 +57,7 @@ public class ItemEdit extends JComponent {
         currentData.description=description.getText();
         currentData.iconName=iconName.getText();
         currentData.questItem=questItem.isSelected();
-        currentData.cost=((Integer)  cost.getValue()).intValue();
+        currentData.cost= (Integer) cost.getValue();
     }
 
     public void setCurrentItem(ItemData data)
